order_service/product/serializers.py
from rest_framework import serializers

from .models import Product

class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = Product
        fields = (
            "id",
            "title",
            "description",
            "dimensions",
            "price",
            "image",
            "get_absolute_url"
        )